Update F - XOR on Grid Path.cpp

+++ AtCoder/AtCoder Beginner Contest/271/F - XOR on Grid Path.cpp
@@ -33,8 +33,8 @@ const int maxn = 30;
 int n;
 int a[maxn][maxn];
 
-vector<int> lef[maxn], rig[maxn];
-void gen1(int x = 1, int y = 1, int cur = a[1][1])
+vector<ll> lef[maxn], rig[maxn];
+void gen1(int x = 1, int y = 1, ll cur = a[1][1])
 {
     if (y > n + 1 - x) return;
     if (y == n + 1 - x)
@@ -46,7 +46,7 @@ void gen1(int x = 1, int y = 1, int cur = a[1][1])
     gen1(x + 1, y, cur ^ a[x + 1][y]);
     gen1(x, y + 1, cur ^ a[x][y + 1]);
 }
-void gen2(int x = n, int y = n, int cur = a[n][n])
+void gen2(int x = n, int y = n, ll cur = a[n][n])
 {
     if (y < n + 1 - x) return;
     if (y == n + 1 - x)
@@ -68,13 +68,13 @@ void solve()
 
     gen1();
     gen2();
-    int res = 0;
+    ll res = 0;
     for (int i = 1; i <= n; i++)
     {
-        map<int, int> mp;
-        for (int x : lef[i])
+        map<ll, int> mp;
+        for (ll x : lef[i])
             mp[x]++;
-        for (int x : rig[i])
+        for (ll x : rig[i])
             res += mp[x];
     }
